pyproject.toml
[project]
name = "shellworks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.26.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "tomli>=2.4.0",
]

[project.scripts]
shellworks = "shellworks.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
